sst-2 bug fixed, notebook fixed
