+++ nemo/collections/nlp/data/datasets/__init__.py
@@ -31,7 +31,10 @@ from nemo.collections.nlp.data.datasets.punctuation_capitalization_dataset impor
     BertPunctuationCapitalizationInferDataset,
 )
 from nemo.collections.nlp.data.datasets.qa_squad_dataset.qa_squad_dataset import SquadDataset
-from nemo.collections.nlp.data.datasets.text_classification_dataset import BertTextClassificationDataset
+from nemo.collections.nlp.data.datasets.text_classification import (
+    BertTextClassificationDataset,
+    TextClassificationDataDesc,
+)
 from nemo.collections.nlp.data.datasets.token_classification_dataset import (
     BertTokenClassificationDataset,
     BertTokenClassificationInferDataset,
